\name{MyRStatLib-package}
\alias{MyRStatLib-package}
\alias{MyRStatLib}
\docType{package}
\title{
\packageTitle{MyRStatLib}
}
\description{
\packageDescription{MyRStatLib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyRStatLib}
\packageIndices{MyRStatLib}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Haruka Sato}

Maintainer: \packageMaintainer{MyRStatLib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
